# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TechSpar",
    "version": "0.4.1"
  },
  "components": {
    "schemas": {
...
      "ValidationError": {
        "type": "object",
        "title": "ValidationError",
        "properties": {
          "loc": {
            "type": "array",
            "title": "Location",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ]
      },
      "HTTPValidationError": {
        "type": "object",
        "title": "HTTPValidationError",
        "properties": {
          "detail": {
            "type": "array",
            "title": "Detail",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      }
    },
...
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "paths": {
...
    "/api/copilot/prep": {
      "post": {
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "jd_text": {
                    "type": "string",
                    "minLength": 1
                  },
                  "company": {
                    "type": "string",
                    "default": ""
                  },
                  "position": {
                    "type": "string",
                    "default": ""
                  }
                },
                "required": [
                  "jd_text"
                ]
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "jd_text": {
                    "type": "string",
                    "minLength": 1
                  },
                  "company": {
                    "type": "string",
                    "default": ""
                  },
                  "position": {
                    "type": "string",
                    "default": ""
                  }
                },
                "required": [
                  "jd_text"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Queued Copilot preparation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prep_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "prep_id"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/copilot/preps": {
      "get": {
        "responses": {
          "200": {
            "description": "Copilot preparations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/copilot/prep/{prep_id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "prep_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Copilot preparation status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "nullable": true
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "prep_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted Copilot preparation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/copilot/prep/{prep_id}/tree": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "prep_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Copilot strategy tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "nullable": true
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  }
}